requirements.txt
Pillow==7.1.2
Pillow==10.4.0
xlrd==1.2.0
